# specs/1-backend-core/contracts/todo-api-openapi.yaml
openapi: 3.0.3
info:
  title: Todo API
  description: REST API for Todo application backend
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
paths:
  /users/{user_id}/tasks:
    get:
      summary: List all tasks for a user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the user
      responses:
        '200':
          description: Successful response with list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a new task for a user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{user_id}/tasks/{task_id}:
    get:
      summary: Get a specific task by ID
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the user
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the task
      responses:
        '200':
          description: Successful response with task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update an existing task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the user
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a task
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the user
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the task
      responses:
        '204':
          description: Task deleted successfully
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{user_id}/tasks/{task_id}/toggle:
    patch:
      summary: Toggle task completion status
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the user
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the task
      responses:
        '200':
          description: Task completion status toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Task:
      type: object
      required:
        - id
        - title
        - completed
        - user_id
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the task
        title:
          type: string
          maxLength: 255
          description: Brief title of the task
        description:
          type: string
          maxLength: 1000
          nullable: true
          description: Detailed description of the task
        completed:
          type: boolean
          description: Flag indicating if the task is completed
        user_id:
          type: string
          format: uuid
          description: Identifier of the user who owns this task
        created_at:
          type: string
          format: date-time
          description: Timestamp when the task was created
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the task was last updated
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 255
          minLength: 1
          description: Brief title of the task
        description:
          type: string
          maxLength: 1000
          nullable: true
          description: Detailed description of the task
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
          minLength: 1
          nullable: true
          description: Brief title of the task
        description:
          type: string
          maxLength: 1000
          nullable: true
          description: Detailed description of the task
        completed:
          type: boolean
          nullable: true
          description: Flag indicating if the task is completed
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Error message
        details:
          type: object
          description: Additional error details (optional)
          nullable: true